Moved dining tags from form_extras to dining tags
