+++ assets/templates/dining_lists/dining_slot_diners.html
@@ -2,6 +2,7 @@
 {% load staticfiles %}
 
 {% load form_extras %}
+{% load dining_tags %}
 
 
 {% block details%}
