+++ assets/templates/dining_lists/snippet_diningslot.html
@@ -1,4 +1,5 @@
 {% load form_extras %}
+{% load dining_tags %}
 
 <div class="col-12 mx-0 my-2 btn d-inline-flex text-left
         {% if interactive %}py-4
